requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI & LLM
mistralai>=0.0.15
httpx>=0.25.0

# Database
supabase>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# External APIs
twilio>=8.0.0
resend>=0.5.0

# Utilities
pydantic>=2.0.0
